+++ src/datumaro/plugins/validators.py
@@ -6,7 +6,7 @@ from copy import deepcopy
 
 import numpy as np
 
-from datumaro.components.annotation import AnnotationType, LabelCategories
+from datumaro.components.annotation import AnnotationType, GroupType, LabelCategories
 from datumaro.components.cli_plugin import CliPlugin
 from datumaro.components.errors import (
     AttributeDefinedButNotFound,
@@ -32,6 +32,8 @@ from datumaro.components.errors import (
 )
 from datumaro.components.validator import Severity, TaskType, Validator
 from datumaro.util import parse_str_enum_value
+
+DEFAULT_LABEL_GROUP = "default"
 
 
 class _TaskValidator(Validator, CliPlugin):
@@ -553,7 +555,7 @@ class ClassificationValidator(_TaskValidator):
         label_name_to_group = {}
         for label_group in label_groups:
             for idx, label_name in enumerate(label_group.labels):
-                if label_group.group_type == 0:
+                if label_group.group_type == GroupType.EXCLUSIVE:
                     label_name_to_group[label_name] = label_group.name
                 else:
                     label_name_to_group[label_name] = label_group.name + f"_{idx}"
@@ -567,7 +569,7 @@ class ClassificationValidator(_TaskValidator):
                 if ann.label in undefined_label_name:
                     continue
                 label_name = label_cat[ann.label].name
-                label_group = label_name_to_group.get(label_name, "")
+                label_group = label_name_to_group.get(label_name, DEFAULT_LABEL_GROUP)
                 if label_group in occupied_groups:
                     stats["items_with_multiple_labels"].append(item_key)
                     break